py/requirements.txt
# requirements.txt
openai>=1.0
streamlit>=1.32
python-dotenv
langchain-openai>=0.0.5
langchain-community>=0.2.0
chromadb>=0.4.24
llama-parse>=0.2.12          # 可选：若用 LlamaParse API
llama-hub>=0.0.43
rank_bm25>=0.2.2
pydantic>=1.10,<3            # 避免 pydantic v3 破坏依赖
tiktoken>=0.6.0
python-docx>=1.1.0
PyPDF2>=3.0.1
streamlit>=1.34.0
pytest>=8.0.0
umap-learn>=0.5.6
plotly>=5.22.0      # Streamlit 原生支持 Plotly
scikit-learn>=1.4.2 # UMAP 依赖